+++ Sources/Parsers/DE/DETimeExpressionParser.swift
@@ -154,7 +154,7 @@ public class DETimeExpressionParser: Parser {
         }
         
         if result.end == nil {
-            result.end = ParsedComponents(components: nil, ref: (try result.start.date))
+            result.end = ParsedComponents(inheritingDateFrom: result.start, ref: ref)
         }
         
         hour = 0
@@ -248,7 +248,7 @@ public class DETimeExpressionParser: Parser {
             }
         }
         
-        if try (result.end!.date).timeIntervalSince1970 < (result.start.date).timeIntervalSince1970 {
+        if result.end!.isDefinitelyBefore(result.start) {
 						try result.end?.shiftCalendarDays(1)
         }